统一矩阵相关类型为int，修正分块与并行实现
本次提交将所有矩阵维度、步长、循环变量等类型从size_t统一为int，简化类型并提升可维护性。修正了多级分块算法和OpenMP并行实现中的类型和边界处理，保证分块正确。调整TOLERANCE宏，matmulFuncs去除未实现版本。同步更新辅助工具函数类型，补充详细中文注释，提升代码可读性。

+++ include/utils.h
@@ -2,12 +2,12 @@
 
 void ClearCache();
 
-void MallocMatrix(const size_t M, const size_t N, const size_t K, size_t& lda, size_t& ldb, size_t& ldc, float*& A, float*& B, float*& C, float*& REF);
+void MallocMatrix(const int M, const int N, const int K, int& lda, int& ldb, int& ldc, float*& A, float*& B, float*& C, float*& REF);
 
 void FreeMatrix(float*& A, float*& B, float*& C, float*& REF);
 
-void InitABCREF(const size_t M, const size_t N, const size_t K, const size_t lda, const size_t ldb, const size_t ldc, float* A, float* B, float* C, float* REF);
+void InitABCREF(const int M, const int N, const int K, const int lda, const int ldb, const int ldc, float* A, float* B, float* C, float* REF);
 
-void PrintABC(const size_t M, const size_t N, const size_t K, const size_t lda, const size_t ldb, const size_t ldc, float* A, float* B, float* C);
+void PrintABC(const int M, const int N, const int K, const int lda, const int ldb, const int ldc, float* A, float* B, float* C);
 
-void CheckResult(const size_t M, const size_t N, const size_t ldc, float* C, float* REF, float tolerance);
+void CheckResult(const int M, const int N, const int ldc, float* C, float* REF, float tolerance);
